+++ spotbugs-tests/src/test/java/edu/umd/cs/findbugs/XDocsBugReporterTest.java
@@ -1,0 +1,30 @@
+package edu.umd.cs.findbugs;
+
+import edu.umd.cs.findbugs.classfile.ClassDescriptor;
+import edu.umd.cs.findbugs.classfile.DescriptorFactory;
+import org.junit.Test;
+
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.MatcherAssert.assertThat;
+
+public class XDocsBugReporterTest {
+    @Test
+    public void reportMissingClassByClassNotFoundException() {
+        Project project = new Project();
+        XDocsBugReporter bugReporter = new XDocsBugReporter(project);
+        ClassNotFoundException ex = new ClassNotFoundException("Class org.example.Foo cannot be resolved.");
+        bugReporter.reportMissingClass(ex);
+        SortedBugCollection bugCollection = (SortedBugCollection) bugReporter.getBugCollection();
+        assertThat(bugCollection.missingClassIterator().next(), is("org.example.Foo"));
+    }
+
+    @Test
+    public void reportMissingClassByClassDescriptor() {
+        Project project = new Project();
+        XDocsBugReporter bugReporter = new XDocsBugReporter(project);
+        ClassDescriptor classDescriptor = DescriptorFactory.createClassDescriptorFromDottedClassName("org.example.Bar");
+        bugReporter.reportMissingClass(classDescriptor);
+        SortedBugCollection bugCollection = (SortedBugCollection) bugReporter.getBugCollection();
+        assertThat(bugCollection.missingClassIterator().next(), is("org.example.Bar"));
+    }
+}